chore(vim-test): use neovim; map esc to get out of insert mode
This allows scrolling within the terminal window!

I still like the idea of using iterm2 tabs, but ideally it runs it in
the tab directly to the right of the current tab (the one I typically
have open for terminal commands), but AI cannot figure out how to do
that in AppleScript.